Migrate federation-aware ClusterAPI (T7.7 / ENG-5041)
WS-M3.2 service migration. Moves the customer-facing federation cluster
surface (capabilities + diagnose + fix + execution-gate binding +
attribute-schema lifecycle) from kamiwaza/cluster.py into
kamiwaza_sdk/services/cluster_federation.py.

Module name disambiguation per design §6.2 T7.7: legacy
kamiwaza_sdk/services/cluster.py covers node/hardware/Ray cluster CRUD;
this new module covers the federation-aware surface. No overlap — kept
separate for now. Investigated whether to merge: the legacy module is
Location/Hardware/Cluster/Node CRUD with completely different return
types; merging would expand T7.7 scope significantly and isn't
load-bearing for the M3.2 unification gate.

Mechanical migration — the M1+ ClusterAPI already used
self._client._request(...). Class inherits from BaseService; references
self.client. Legacy kamiwaza/cluster.py becomes a re-export.

26 legacy cluster + diagnose tests pass via the re-export.

References: ENG-5041 (T7.7), design v0.3.7 §4.2.11 §6.2 T7.7.
